Had to remove the check for spaces in the input
The check for if there was space in the input string was causing the output to be incorrect. It was inverting signs and causing the terms not to combine right. Also the output warning of invalid character would display when any polynomial was inputted regardless if it was correct or not.